Add files via upload
This YAML script creates an S3 bucket and configures it to host a static website. It also creates a DynamoDB table with a simple primary key based on a product ID attribute.

To use this template, you can save it as a .yaml file and deploy it using AWS CloudFormation. However, this is just the infrastructure part, and you'll need to develop the actual e-commerce application separately.